+++ CHANGELOG.md
@@ -8,3 +8,7 @@
 - Created IMC (Inter Module Communication)
 - Removed unused fields.
 - Created yaml_encoding to encode json into yaml files
+
+## 0.2.1
+- Added support for initialize only actives module
+- Commented topological sort